Ability to cancel an ongoing read-aloud, both from a notification on the phone and on the Wear device.
This necessitated changing JorSayReader from an IntentService to a Service. The crucial benefit is we now decide when to stop - whereas an IntentService onDestroy() was called as soon as onHandleIntent() was done, we can now call stopSelf() when TTS is done - either because reading is complete, or canceled or there was an error.

Testing done:
1. 'cancel read aloud' notification is posted when reading aloud starts
2. Ongoing reading can be canceled from both phone and Wear
3. Cancel icon looks visually fine on both phone and watch
4. No regressions. Notifications are still read aloud reliably.