Snooze ongoing 'read aloud'
This provides an alternative implementation for the requirement of stopping an ongoing read-aloud. JorSayReader was changed to a Service from an IntentService while implementing 'cancel read aloud', but major problems remained - TTS was still being initialized for every read-aloud command, there was a new UtteranceProgressListener for every TTS instance, and stopping read aloud did not do it for those TTS instances that were already created and scheduled. All of that is fixed now. JorSayReader now has ONE TTS instance and has ONE UtteranceProgressListener associated with it. (Android has a per-TTS instance queue of utterances, and stopping a TTS instance causes that queue to be flushed. However, and this is important, this does not affect any other tts instances from the same or other apps.) Snoozing is now moved to a separate class (and package). The status in the main screen is now aware of the snooze and also provides a way to cancel a snooze and resume reading from next incoming notification.

Testing done:
1. Reading aloud can be snoozed while it is going on
2. Different snooze duration work as expected
3. Snoozing works from the notification on Wear
4. Decrementing snooze duration is displayed on the main screen
5. Snooze can be canceled from the main screen